fix for items.
Note: It doesnt appear to change anything in the parcelware web items number, regardless of how many there are. This is also just simple implementations. As there are 3 ways to have products shipped, and it would be a complex endeavor to allow the user to specify what goes in what boxes ( or if they all go in 1 box ).